Added test helper for end to end environment config.
